7.3-7.5 Accounting views for marketing agency, marketing agent, recruitment agent
- 9 new tests (AccountingAgentTest) with tenant scoping
- AccountingController::marketingAgency(), marketingAgent(), recruitmentAgent()
- Views: accounting/marketing-agency.blade.php, marketing-agent.blade.php, recruitment-agent.blade.php
- Routes: accounting.marketing-agency, accounting.marketing-agent, accounting.recruitment-agent
- 2.13: status_transitions migration with FK constraints to status_codes.code
- Todo file: marked 7.1-7.6, 8.1-8.2, 8.8 as done

+++ database/migrations/2026_06_16_120000_create_status_transitions_table.php
@@ -11,8 +11,8 @@ return new class extends Migration
         if (!Schema::hasTable('status_transitions')) {
             Schema::create('status_transitions', function (Blueprint $table) {
                 $table->id();
-                $table->integer('from_code');
-                $table->integer('to_code');
+                $table->unsignedSmallInteger('from_code');
+                $table->unsignedSmallInteger('to_code');
                 $table->boolean('is_active')->default(true);
                 $table->timestamps();
